Redesigned the application. Changed the data loading class. Added the "UIImageView" extension. Removed "SceneDelegate". Launched with "AppDelegate".
