+++ Cartoons For Kids_image/CartoonsForKidsList/CartoonsForKids_Router.swift
@@ -11,7 +11,7 @@ protocol CartoonsList_Router_Protocol {
 	var entry: CartoonsForKids_ViewController? {get}
 	static func startExecution() -> CartoonsList_Router_Protocol
 
-	func gotoDetailView(cartoon: Cartoon)
+	func gotoDetailView(cartoon: CartoonModel)
 }
 
 
@@ -39,7 +39,7 @@ class CartoonsForKids_Router: CartoonsList_Router_Protocol{
 	 
  }
  
- func gotoDetailView(cartoon: Cartoon) {
+ func gotoDetailView(cartoon: CartoonModel) {
 	 
  }
