--- Cartoons For Kids_image/ViewController.swift
@@ -1,18 +1,0 @@
-//
-//  ViewController.swift
-//  Cartoons For Kids_image
-//
-//  Created by Николай Гринько on 27.08.2024.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-	override func viewDidLoad() {
-		super.viewDidLoad()
-		view.backgroundColor = .red
-	}
-	
-}
-
